change entity "type" field to "entity_type"
This allows us to use the discriminator entity_type to determine the
type of entity to create when deserializing IOC yaml. BUT does not
clash with the type field used for Entity Args.

This will allow us to switch to using union of types for Args and
therefore have them defined as their type or a Jinja template.